Fix IPv6 Socket (#228)
* Change

* Bump version

* update nugets

* update nugets

* Alcala requirement

* Extending BinaryResponseAsync

* Codacy

* Update

* Test

* Fixing issues with connections

* Update README.md

* Add Response Extensions

* Add FileInfo as option to Fallback Module

* Rename extension method

* Minor

* Cleanup

* Update appveyor.yml

* Fix

* Ipv6 tests[

* My Bad

* Test

* Test without ipv6

* Refactoring

* Check

* Changes to Style

* Fix build

* Fix partial

* Fixing some unit tests

* My bad

* Fix websocket

+++ src/Unosquare.Labs.EmbedIO/System.Net/WebSocketFrame.cs
@@ -102,12 +102,12 @@
             }
             else if (len < 0x010000)
             {
-                PayloadLength = (byte)126;
+                PayloadLength = 126;
                 ExtendedPayloadLength = ((ushort)len).ToByteArray(Endianness.Big);
             }
             else
             {
-                PayloadLength = (byte)127;
+                PayloadLength = 127;
                 ExtendedPayloadLength = len.ToByteArray(Endianness.Big);
             }
